PEP 8 Code + MD5 Hashes in Gen JSON
Changed the script to use PEP 8 compliant code with 4 space
indentation and other fixes.

A subprocess.call("git pull -q") to update our repo from github.

It contains MD5 Hashes of all files (which will be used to check if
a file has been modified or not.)

All files are scanned for plugin data and not just __init__.py

The JSON output is now dumped to a file rather than the screen.